[fidl] Remove usages of protocol aliases
Using protocols as an alias is not allowed in the new syntax, where
protocols are not treated as types. In the old syntax, the
alias:

```
alias TypeOrProtocol = MyProtocol;
```

can have two meanings: the protocol MyProtocol, e.g. in:

```
alias MyProtocolServerEnd = request<TypeOrProtocol>;
```

or the client end of MyProtocol, e.g. in:

```
alias MyProtocolClientEnd = TypeOrProtocol;
```

Such an ambiguity does not exist in the new syntax since usage
as a protocol is not allowed. This CL removes instances of
using an alias as a protocol in preparation for conversion to the
new syntax.

Since the migration is happening soon and this is not a commonly
used feature, it doesn't seem worth the effort to prevent new usages
by updating fidlc.

Test: fx test fidlc_golden_test
Change-Id: Ib85f350bb1486a6ab7c89c4f116675c7083ef18a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/536940
Commit-Queue: Felix Zhu <fcz@google.com>
Reviewed-by: Alex Zaslavsky <azaslavsky@google.com>