[test][debug] Signal when a debug VMO may be read
This is the first of two changes to add explicit signals to DebugData
processing.
This change updates the DebugData protocol and makes the minimal
changes needed to successfully build. This adds an empty protocol to
DebugData.Publish, and updates clients to leak the client end of the
channel. The channel is implicitly dropped when the client process
terminates, which signals DebugData.
Note that this does not update test manager's DebugData implementation
yet. The server side will be updated in the next change.

Bug: 75686
Testing: CI dryrun, existing tests

Change-Id: I77ed69b94d8eeadf1998d03fe0d4a88d4721882d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/532986
Commit-Queue: Satsuki Ueno <satsukiu@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Rich Kadel <richkadel@google.com>
Reviewed-by: Ankur Mittal <anmittal@google.com>
Reviewed-by: Petr Hosek <phosek@google.com>
API-Review: Roland McGrath <mcgrathr@google.com>