[e1000] Use driver protocol methods rather than manual scanning
The e1000 driver scans the BAR registers in the function's config
space to figure out where its IO bar is. This isn't disallowed, but
it's not recommended and may be locked down in the future. The method
now uses PciGetBar to obtain the BARs and check if they are the IO
BAR.

Change-Id: Ib576b7b65c7b1d131fc0aed98f444ab2513627d7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/540264
Commit-Queue: Christopher Anderson <cja@google.com>
Reviewed-by: John Grossman <johngro@google.com>