[flatland] Replace Scale with ImageSize
Remove the ability to set scale on a flatland transform
and instead directly set the size of an image-content
which is local to that image and must have integer
width and height values.

Testing: Existing tests that are modified to accommodate
the API change.

Change-Id: I3ff24f96fe635aec1ec84ea7a7c481260291a5fb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/534821
Commit-Queue: Chris Fontas <cfontas@google.com>
Reviewed-by: Emircan Uysaler <emircan@google.com>
Reviewed-by: John Bauman <jbauman@google.com>
Reviewed-by: David Reveman <reveman@google.com>
API-Review: John Bauman <jbauman@google.com>