[fidlfmt] Add DivisibleSpanSequence to the formatter
As of the previous change, the formatter had no way to describe lines
that should be wrapped if they grow overlong, as DivisibleSpanSequence
was not yet implemented.  This change implements it, allowing us to
properly test our first wrapping use case: `use foo as bar;`, with the
wrapping happening before the `as` token if needed.

Additionally, this change fixes a few edge case bugs from the previous
one.

Test: fidldev test fidlc
Bug: 73507
Change-Id: I55a2bf9781f6c451691394ff4c39b75f855e0eec
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/535543
Commit-Queue: Alex Zaslavsky <azaslavsky@google.com>
Reviewed-by: Felix Zhu <fcz@google.com>