[system-update-checker] Make current channel immutable.
This CL modifies the behaviour of system-update-checker to be closer to
that of omaha-client. It removes use of current_channel.json and instead
determines current_channel once at startup based on either
channel-in-vbmeta or rewrite rules.

Current channel is now constant for each boot of the system on all
products.

Bug: 59881
Change-Id: Id3b27856f5c1066eadf022fcb609a6a30d846f61
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/511718
Commit-Queue: Simon Shields <simonshields@google.com>
Reviewed-by: Dan Johnson <computerdruid@google.com>