[fxfs] A/B superblock
Adds a B copy of the superblock which is stored contiguously after the
current superblock (at a known offset).

This doubles the amount of reserved space we have for superblocks from
512K -> 1M; we might want to tune that down later, but it is fine for
now.

The superblocks have an increasing generation number which is used to
determine which one is newest.

I've also added a few basic fields (magic, version) to the superblock.
Currently the minor_version field isn't used.

Tested: Unit.
Change-Id: Ia4e5ca5055cb5c45e0cf159f969f0f5098387c59
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/537743
Commit-Queue: James Sullivan <jfsulliv@google.com>
Reviewed-by: Chris Suter <csuter@google.com>