[devshell] Don't copy unchanged files from remote
This CL adds the `--times` parameter to the rsync invocation for getting
files from the remote build directory. By default, unless --checksum is
passed, rsync uses the modification time + file size to determine if a
file needs to be transferred. If modification times are not preserved
because `--times` is not used, the files will always look different to
rsync and be transferred.

Bug: 78132
Test: manually invoke `fx flash-remote` twice, ensure no transfer
happens, but a transfer does happen after a new remote build

Change-Id: I506197b23ee85b0e02821c2a42220ab6a8c1d04b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/538811
Reviewed-by: Amit Uttamchandani <amituttam@google.com>
Reviewed-by: Renato Mangini Dias <mangini@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Marc Khouri <mnck@google.com>