[graphis][vkcopy] Validate only last copy operation in benchmark
Mapping and validating the copy operation may take a nontrivial amount
of time, so move it out of the measured portion of the test.

Also add a flush before the copy and an invalidate after, since we don't
request a host coherent memory type. We should be doing a pipeline
barrier to VK_ACCESS_HOST_READ_BIT before the invalidate, but that can
happen in a later patch.

Change-Id: I9c08c68500cc9e5395df26e0fb29c15c2a92a835
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/523103
Commit-Queue: John Bauman <jbauman@google.com>
Fuchsia-Auto-Submit: John Bauman <jbauman@google.com>
Reviewed-by: David Reveman <reveman@google.com>
Reviewed-by: Craig Stout <cstout@google.com>