[lockdep] Fix out of bounds reads.
This CL changes the way the LockDependencySet Iterator works.
Previously, the operator++ would check if the iteration index was valid
before increasing, but after increasing it it could become invalid, for
example, if it was equal to kMaxLockDependencies - 1.

This would cause an out-of-bounds read inside the set.list_ array, which
was detected by KASAN.

The new version, increases the index before checking against
kMaxLockDependencies. The behavior is identical for all cases, except
for when index == kMaxLockDependencies. But that's ok, because you
shouldn't advance an iterator that is at the end.

Bug: 78333
Change-Id: I3a05fc87795e67efba3a837d497dc98aa90650d4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/540584
Reviewed-by: Corey Tabaka <eieio@google.com>
Commit-Queue: Marco Vanotti <mvanotti@google.com>