[brcmfmac] Prevent spurious log about bus state
When the driver is initialized it will set the bus state to down.
Sometimes the variable used to keep track of this state would already be
initialized to that state when it was allocated. This resulted in an
error message about already being in that state. Since there is no
meaningful previous state it's functionally equivalent to just set the
variable right away without calling the helper function to do it.

Bug: 78276
Test: Build, boot and verify connectivity, only affects initialization
Change-Id: Ifbd1a66f527cac94966ca152c3270447f7e02a72
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/539876
Reviewed-by: Joshua Conner <joshconner@google.com>
Commit-Queue: Bjoern Johansson <bjoernj@google.com>