[assembly] Truncate package files before writing them
In the case where a new package is smaller than the previous package, we
were not truncating the length of the file to match the new size,
leaving excess bytes at the end. We now truncate the file to a size of 0
before writing to them.

x = old package
y = new package

[ xxxxxxxxxxxxxxxxxxxxxxx ]
[ yyyyyyyyyyyyyyyyyyxxxxx ]

This was causing the generated files to be invalid packages, and was
easy caught by switch products (via fx set) in the same out directory.

Change-Id: I142cf536fd5c2d1d444c6c74f2a5f033b1a7a43f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/540884
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Aidan Wolter <awolter@google.com>