[memalloc] Add algorithm for finding normalized RAM ranges
In early physboot bootstrapping, we need scratch memory for the
bookkeeping of memory itself. Today we accomplish this by allocating
a sizable, guestimated chunk of the .bss for this purpose. This change,
however, introduces a slick means of efficiently inspecting the span of
arbitrarily-overlapping memory ranges provided in the ZBI to find all of
the encoded, maximally contiguous ranges of RAM within, from which we
will easily be able to find our scratch.

Bug: 77359
Change-Id: I72ee90f3170b1e6f43cf236dc9d4c3b5274d70c2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/538782
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: David Greenaway <dgreenaway@google.com>