[debugger] Start decoding some return values.
Add some infrastructure (currently unused outside of tests) to find
what the return value for a function call was. This will be hooked up in
a later pass to display results during certain stepping operations.

This adds a new ABI class on the ArchInfo structure. Previously the
ArchInfo was only set when there was a connection and we knew the
architecture. Now since this is used in more places, an "unknown"
default architecture is set to avoid lots of null checks. The bulk of
this change is setting up this context properly.

Currently supported are basic types on x64 returned in a single
register, and complex structs returned by value. ARM support lags and
will be filled in more in a later pass.

Bug: 5720
Change-Id: I10904e7db1ab8ab645d559c3d26b2bc4b25d6afd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/539361
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Dangyi Liu <dangyi@google.com>