[botanist] Implement device flash support
Teach botanist how to use flash.sh to flash a target device instead of
paving it. Note that this change does not actually enable the usage of
flashing, as that is gated on foundation flipping the bit to enable
fastboot on hardware. Additionally, this solution by itself is
insufficient, as flash.sh does not support in memory booting of bringup
or ZBI test images. Still, this is a good starting point.

Bug: 78321
Change-Id: I07919a1f704832c76f0187975d3520d6d05e605d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/540481
Fuchsia-Auto-Submit: Anirudh Mathukumilli <rudymathu@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>