[debugger] Show return values after "finish"
Shows the return values (when decodable) when a "finish" command is
executed.

In addition to the current limitations of GetReturnValue, this currently
has some issues with ordering. If there are inline frames after the
physical frame being finished, the thread will continue after the
physical frame "finish" and the registers will be invalid. This
asynchronous problem also means the return value is printed after the
code listing for the next command (which seems off).

To remedy this in a future pass, the callback will need the ability to
suspend the step operations until it has completed.

Bug: 5720
Change-Id: I6e1e34ed3d98353546cb532e05eba7c5f8f9ad18
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/540886
Reviewed-by: Dangyi Liu <dangyi@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>