[bt][gap] Require bond data with IRK to have identity address
Assert that LE bond data passed to PeerCache::{AddBondedPeer,
StoreLowEnergyBond} that has one of {IRK, identity address} has both.

When restoring LE bond data in HostServer, set the identity address to the
peer address if an IRK is specified.

This fixes a bug where restored bonds that specified an IRK did not have
their identity resolving list entry removed when
PeerCache::RemoveDisconnectedPeer was called. These stale entries caused
later attempts to insert a peer with an RPA that was resolvable with a
stale entry to fail (even though PeerCache::FindByAddress would not
return a peer).

Fixed: 78142
Test: fx test //src/connectivity/bluetooth/core
Change-Id: Ic8e5b92e183337615592e41b08fe24fda3c419d7