[fint] Add fvdl_intree_test to list of never affected tests
This is an e2e test that depends on the fvm so it's technically
affected (at least according to the build graph) by a large fraction of
fuchsia.git changes despite the test not actually being sensitive to
most such changes. So it's not useful to automatically multiply it on
those changes, and considering affected also makes it more likely that a
build will exceed the maximum of 3 affected tests allowed before no
affected tests will be multiplied in a build.

Bug: 68434
Change-Id: I973f5f6c316ef9783fcead14583565ca6aa71e4a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/541181
Reviewed-by: Yuan Zhi <yuanzhi@google.com>
Reviewed-by: Marc-Antoine Ruel <maruel@google.com>