[zxio][tests] Refactor file/directory server types
This updates the file and directory FIDL server implementations for
testing to use a common test base built on top of the LLCPP-generated
test base types. This also updates the base classes to emit explicit
failures on unexpected messages with the name of the message in the
failure message.

Bug: 43267
Change-Id: I32f58feff0331f8f95183725d2738a5da52236ba
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/541102
Commit-Queue: James Robinson <jamesr@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>