[rbe] prototype compile wrapper for rustc on RBE
build/rbe/rustc-remote-wrapper.sh

The intent is for this script to be used as a command prefix
in rustc tool invocations, using the tool-agnostic
fuchsia-rbe-action.sh.
This script scans the rust compile command for a variety of inputs
and outputs (including depfile).

Input sources are taken from a locally generated depfile.
We use -Zbinary-dep-depinfo to get the transitive set of dependencies
needed to be uploaded.

Uses ldd to find related shared libraries for the rustc binary.

Rust standard libraries are listed in dep-info.

--inputs,--output_files paths are normalized relative to $project_root
which is $FUCHSIA_DIR, in practice.

Missing: dependent clang toolchain binaries

Checked with shellcheck.

Tested: (manual invocation)
  [from output dir, command from Ninja build rule for libsyn.rlib]
  ../../build/rbe/rustc-remote-wrapper.sh -- \
  ../../prebuilt/third_party/rust/linux-x64/bin/rustc \
  --color=always \
  --crate-name syn \
  ../../third_party/rust_crates/vendor/syn/src/lib.rs \
  --crate-type rlib \
  --emit=dep-info=obj/third_party/rust_crates/libsyn-59df571bc0ccf4fd.rlib.d,link \
  ...

Change-Id: Ic0b1902dd64683e62b2634cfe817f5e5441f994a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/537821
Commit-Queue: David Fang <fangism@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Reviewed-by: Tyler Mandry <tmandry@google.com>
Reviewed-by: David Turner <digit@google.com>