[debugger] Allow porting of stdio
Unify the APIs of BufferedZxSocket and BufferedFD which were almost the
same but differed in some details (FD started watching on Init, ZX
didn't). Remove the Init phase which made construction more complicated
and just assert that the handles passed to the constructor are valid.
The common API from these classes is separated into a new BufferedStream
class.

Add a typedef in debug_agent for the the platform type to use for stdio
streaming. Use this new BufferedStdioHandle instead of BufferedZxSocket.

Change-Id: Ib6c24fe67b10454697752b71bad03c9cd7d1771d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/540302
Reviewed-by: Dangyi Liu <dangyi@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>