[wlan][policy] clean config_manager code
This removes lower_valid_security, which was added to support recording
connect results using the security type of the actual connection, which
could have been upgraded from the saved config's security. However this
is no longer needed because network selection and the state machine pipe
through the security of the actual network config to record the result
of the connection attempt.

Bug: 77060
test: fx test wlancfg-tests (no functionality added in this change)

Change-Id: Ibce228075f2d23fa63052e8f2944cd367e4f6992
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/541501
Commit-Queue: Naomi Mccracken <nmccracken@google.com>
Reviewed-by: Marc Khouri <mnck@google.com>