[wlan][policy] Remove expected state machine error
This change modifies the exit reason when we exceed the number of
connection attempts in the state machine. This is an expected condition,
so we shouldn't treat it as an error.
Returning an `Err()` here causes an error message to be logged.

This boils down to a logging-only change, because the contents of the
`ExitReason` is only used for logging, not flow control:
https://cs.opensource.google/fuchsia/fuchsia/+/master:src/connectivity/wlan/wlancfg/src/client/state_machine.rs;l=151-158;drc=80903f7b7ec4ed2a9b21b17c434f2ce48dae403d

Bug: 78486
Test: fx test wlancfg-tests
Change-Id: If9eeab8d7c072fa8e1492756996343ce17955791
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/541862
Fuchsia-Auto-Submit: Marc Khouri <mnck@google.com>
Reviewed-by: Naomi Mccracken <nmccracken@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>