[driver-manager] Fix behaviour of ChildListIterator for composites.
Previously, ChildListIterator would look at the device's parent's
fragments to figure out what composite children a device has. This
resulted in composite devices being shown as children of every device
which was their parent's sibling.

For instance, take a simple example topology:

        /---> child
device -----> fragment -----> composite device

Iterating over child's children would show that it had "composite
device" as a child! This is because the iterator wasn't checking that:

a) The device was a fragment before iterating over composite child
   devices (since only fragments can have composites as their child)
b) The device whose children we were iterating over was the device to which
   the fragment was bound (since the parent may have more than one
   fragment child).

Test: `dm dump` no longer shows composites in the wrong places.
Fixed: 75946
Change-Id: I4036e89d153017b7e7c4cfb2de456b1213eb73e6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/540061
Commit-Queue: Simon Shields <simonshields@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>