[fxfs] Purge old layer files after compaction
This CL also fixes a deadlock that could be triggered when zeroing the
journal.  It was taking the same lock that maybe_extend_journal_file was
taking, so I've moved things around so that extending and zeroing the
journal file is in the same place.

Change-Id: I772e3236a40b6d3011cc11476169105d1f6d06c3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/541163
Fuchsia-Auto-Submit: Chris Suter <csuter@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>