[fxfs] Wrap the filesystem with something that will close
Wrap the filesystem in something such that when it is dropped, it
automatically calls close on the filesystem if it hasn't already been
called.

This will go some way to stopping crashes in tests where close isn't
called.

To fix this properly though, we need to flip the strong/weak references
around so that the strong relationship is from child to parent and then
we also need to add strong references to any object stores that have
data that needs flushing.

Change-Id: I0432f8843e0c824f4a89a3063b50b897abd067f7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/541301
Commit-Queue: Chris Suter <csuter@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>