[sys] Make *ServiceDecl.source singular.
Prior to this change, the native versions of OfferServiceDecl and
ExposeServiceDecl had a Vec of sources, unlike
{Offer,Expose}ProtocolDecl and friends. In theory, this paved the way
for a FIDL service to be constructed from multiple sources, but that
functionality was never implemented and isn't currently planned.

Since the FIDL versions of OfferServiceDecl and ExposeServiceDecl only
had single source, it was not possible to do a 1:1 conversion between
the native and FIDL versions of these types, which is exactly what I
need to do as part of Dynamic Offers.

We can revisit this functionality in the future, but if we do, the
type conversion layer is probably not where the service aggregation
belongs.

Bug: fxbug.dev/71881

Change-Id: Ia5d118ee25e33ccfb3677d2987758e512ff9dcf0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/542083
Reviewed-by: Adam Lesinski <adamlesinski@google.com>
Commit-Queue: Hunter Freyer <hjfreyer@google.com>