[gidl][rust] Add test to detect regressions in Rust GIDL build with no tests
Currently, there are often build issues with GIDL in various bindings
when tests are disabled. This can make it hard to disable tests for
debugging. This CL adds a test that always builds rust GIDL output with
an empty GIDL input file.

The idea is that this would be done for other languages as well. I'm
starting with rust to test out the idea so that work is not wasted doing
other bindings the wrong way if there are issues with this approach.

Change-Id: Ia7a2368f0ff396299f0e027dc70d7567dc793d1d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/532001
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>
Reviewed-by: Mitchell Kember <mkember@google.com>
Reviewed-by: Pascal-Louis Perez <pascallouis@google.com>