[vfs] Fix blobfs integration tests in new pager.
Fix the remaining blobfs integration tests while running under the new
pager (currently defaulted to "off").

The first problem was that Blobfs teardown needed a way to force any
remaining live Blob objects to release their resources and release their
internal pager reference. This necessitated returning the reference from
FreePagedVmo so the caller can control the deletion time if necessary.

The second problem was that the watch on the VMO in PagedVmo wasn't
actually getting canceled. You have to explicitly call Cancel() and not
just clear the object. It also requires that the watch object be cleared
before the VMO is deleted or there will be assertions about freeing
watches on invalid handles.

Bug: 51111
Change-Id: Ib3527ca77bec9cd8f57572656c37eca8ff588157
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/541828
Reviewed-by: Martin Lindsay <mlindsay@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>