[flatland] Errors close channel after returning OnPresentProcessed
This CL changes the way errors are issued in flatland:
- Errors are now returned as a param of OnPresentProcessed instead of
Present callback. This means one less wakeup for the client to check
the results.
- All errors are now unrecoverable and close the channel. Since
illegal operations are client's fault and may result in updates that
are not clear to the client or Scenic, we decided to treat them all
as fatal.

Fixed: 56869, 36166
Change-Id: I54a2b160667734783f476fbd86a04a99e49dc7f6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/540601
Commit-Queue: Emircan Uysaler <emircan@google.com>
Reviewed-by: John Bauman <jbauman@google.com>
Reviewed-by: Josh Gargus <jjosh@google.com>
API-Review: John Bauman <jbauman@google.com>