[media][audio] Require effect "name" in json config
The EffectsController FIDL service applies effect updates based on a
unique instance "name" for each effect. Currently, effect names are not
required in the output pipeline config, and this can break the
EffectsController service. To prevent this failure, this CL requires
each effect in the output pipeline config to specify a unique instance
"name".

Test: fx test audio_core_unittests
Change-Id: I31c2cf77a70c64425f448a8bb88795deb54e36d9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/542384
Commit-Queue: Payton Bradsky <pbradsky@google.com>
Reviewed-by: Tom Bergan <tombergan@google.com>