[goldfish-sync] Fix a heap use-after-free error.
In sync device SyncTimeline::CreateFence() function, if a
fence is deleted from an inactive SyncTimeline (so the timeline
is not in the SyncDevice anymore), deletion of the fence may
cause the SyncTimeline object to be deleted as well, which could
cause an heap use-after-free error when releasing the lock
in the callback.

This change fixes this by using a RefPtr to the SyncTimeline
in the callback instead.

TEST=added a minimal test case SyncDeviceTest.TimelineDestroyedAfterFenceClosed
     to goldfish-sync-device-test to cover this case.
Fixed: 73217

Change-Id: I20ac7e635ce8f7fa4370336bb3fb4f49b701d066
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/542381
Reviewed-by: David Reveman <reveman@google.com>
Reviewed-by: Marco Vanotti <mvanotti@google.com>
Commit-Queue: Yilong Li <liyl@google.com>