[build] Refactor renamed_binary() implementation.
This CL changes the distribution manifest format to ensure that
renamed_binary() targets do not preserve the original install
location by default, as this is generally undesirable.

Doing so requires introducing new 'renamed entries' and
'copy entries' in the partial distribution manifest format,
to deal with the fact that build variants place binaries
under $BUILD_DIR/x64-variant/foo instead of $BUILD_DIR/foo.

An example is given in manifest_formats.md to better explain
the situation.

As explained in the updated documentation for renamed_binary(),
while it is possible to use several renamed_binary() targets
with the same source, it is not possible to mix any of them
with a resource() that uses the same source. Our processing
scripts (see distribution_manifest.py) detects these cases and
prints an error message.

+ Modify gub.gni and gsb.gni to use renamed_binary() instead
  of resource() to install the renamed binary, as this causes
  a build-time conflict.

+ Ensure //build/images/assembly/BUILD.gn uses
  distribution_manifest() instead of doing the metadata walk
  which generates a partial distribution manifest, whose
  format is no longer supported by the assembly ffx plugin.

Change-Id: I80842d2d80e935a126304bebaeed27b4d18a4f6c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/538141
Reviewed-by: Shai Barack <shayba@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Reviewed-by: Dan Johnson <computerdruid@google.com>
Reviewed-by: David Fang <fangism@google.com>
Commit-Queue: David Turner <digit@google.com>