[artifactory] Use relative paths to the checkout for host tests.
As described in the attached bug, some host tests want to have
dependencies on prebuilts which are in the checkout root. This was
causing artifactory to try to append a path starting with "../.." to the
upload namespace (because the host test paths are relative to the build
dir). Setting the paths relative to the checkout root should ensure that
all the host tests and their deps get uploaded under the expected
namespace.

Bug: 77782
Change-Id: If63f3172d75b37a0a4896f99f7c8c1fd4efb8aa5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/540262
Fuchsia-Auto-Submit: Ina Huh <ihuh@google.com>
Reviewed-by: Jay Zhuang <jayzhuang@google.com>
Reviewed-by: Anthony Fandrianto <atyfto@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>