[audio] Add turn_on_delay to driver APIs
We add turn_on_delay to both RingBuffer and Codec protocol.
The RingBuffer to provide it to clients of the StreamConfig or
Dai protocols (for instance audio_core or bluetooth) and Codec
to extract the information from hardware audio codecs and provide
it to glue drivers that may use it to inform their clients.

Test: aml-g12-tdm-test
Change-Id: I4ffc0e085da292aa0a8a97560926e4a3f881048d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/535242
Reviewed-by: Marie Janssen 💖 <jamuraa@google.com>
API-Review: Marie Janssen 💖 <jamuraa@google.com>
Commit-Queue: Andres Oportus <andresoportus@google.com>