[debugger] Add templatized symbol conversions.
Add templaized "As" functions to the Symbol class for accessing its
derived classes (our manual RTTI implementation). My plan is to replace
all callers with these variants and then we can delete the non-const
manual overrides in a later pass (so the net header complexity increase
should be minimal).

With these templatized versions, we can create templatized versions of
some of the common conversion functions. This reduces boilerplate and is
less error-prone at the call sites.

Add a version of GetConcreteType which takes a LazySymbol object, saving
the callers from having to do a conversion and type check before
calling.

Change-Id: Id953ccfe218e63ffe753681c18e450e61626d3f2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/542644
Reviewed-by: Dangyi Liu <dangyi@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>