[fidlfmt] Fix new formatter crash
Previously, the formatter made the incorrect assumption that, in the
case where a protocol has both `compose` and method clauses, the first
instance of the former would always precede the first instance of the
latter.  This is not necessarily true, so we now check to see which of
the two comes first.

Test: fidldev test fidlc
Bug: 73507
Fixed: 78688
Change-Id: Ib30e55959d2c1ea14c3298a9a11ec646a8ff37b0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/542823
Commit-Queue: Alex Zaslavsky <azaslavsky@google.com>
Reviewed-by: Felix Zhu <fcz@google.com>