[fvdl] Add retry when fvdl exited with unexpected fails.
If fvdl fails with exit code 1, it indicates a logic error, and the test
should fail. If fvdl exits with code > 1, it means either aemu crashed
or too slow to start or ssh connection is not up; when this happens, we'll retry running
fvdl.

Bug: 78190
Change-Id: Idc19e61323a82c97d780a1e3f4169d8bd0263aab
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/541141
Reviewed-by: Marc-Antoine Ruel <maruel@google.com>
Commit-Queue: Yuan Zhi <yuanzhi@google.com>