[debug] add handler class for the debug automation
The automation handler is a class that sits between the executor and
the debugged thread. It makes it easier to test the automation
instruction executor, and later it will make testing the automation
process in the debugged thread easier as well.

Change-Id: I461e5757e56c467b7c348de4ca1f892c3405e17b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/541001
Reviewed-by: Brett Wilson <brettw@google.com>
Reviewed-by: Vincent Belliard <vbelliard@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Michael Jones <michaelrj@google.com>