[netdevice] Add migration driver stub
Add an initial stubbed out netdevice migration driver. The driver
implements EthernetIfcProtocol, so that it can be the child of an
ethernet device, and NetworkDeviceImplProtocol, so that it can be the
parent of the netdevice driver. This migration driver is intended to
stand in for the generic Ethernet driver such that concrete ethernet
devices, like WLAN, can have a netdevice descendent driver without the
concrete device driver having to be rewritten in terms of
fuchsia.hardware.network.device banjo.

Bug: 42640
Tested: fx test netdevice-migration-tests

Change-Id: I0f9f01f536b195eb704d5065fd2b80dec476c9d7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/540402
Commit-Queue: Amanda Tait <atait@google.com>
Fuchsia-Auto-Submit: Amanda Tait <atait@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>