[memalloc] Make MemRangeStream work with an arbitrary number of arrays
This will allows us to populate an allocator with an arbitrary number of
range arrays. It will be more convenient to pass in three arrays, as
opposed to the previous two: at that point, might as well make it `n`.

Bug: 77359
Change-Id: I1d9830dfc6fe82c48c07c64fa2a46d101db4d8a3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/542302
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>