[fidl-lint] Convert fidl-lint tests
This CL updates all lint tests to first convert to the new syntax, then
run the test on the converted FIDL.  Since a lot of the inputs aren't
actually valid FIDL (they're templates), extra work is needed to allow
converting this to the new syntax.  In particular, we do a
transformation on the template FIDL, convert, then undo that
transformation to obtain new syntax FIDL templates.

Test: test fidldev fidlc
Bug: 72121
Change-Id: I4564ae57b8ae7a00697e99eb410079964de56f9e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/542602
Commit-Queue: Alex Zaslavsky <azaslavsky@google.com>
Reviewed-by: Felix Zhu <fcz@google.com>