[pkg-cache] Unify dynamic/retained package indices
The pkg-cache needs to track 2 mutable indices of packages, both the
dynamic index and the retained package index. While these indices need
to be tracked independently, the pkg-cache often needs to operate
against one or the other (or both), and it would be inconvenient to pass
around references to both and decide which index to operate against at
each call site. So, this change implements and uses a new PackageIndex
type that composes a DynamicIndex and RetainedIndex, exposing a similar
API surface as the inner indices that delegates to the appropriate inner
index.

Fixed: 77361
Change-Id: Ic63133d365905114330675dffab1a5e081823f0e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/539875
Reviewed-by: Dan Johnson <computerdruid@google.com>
Reviewed-by: Viktar Donich <viktard@google.com>
Commit-Queue: Kevin Wells <kevinwells@google.com>