[blobfs] Fix FragmentationMetrics test
Add a mutex to the test's Logger to prevent races between the flushing
thread and the test thread.

Accumulate the histogram metrics instead of resetting on each new log.
If a flush happened while the metric was being calculated then the test
would drop part of the metric and the test would fail.

Fixed: 78257
Multiply: blobfs-unittest: 50
Change-Id: I2b7fddbc78c41f0c975628f99bcf843ebbb56abc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/542082
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Chris Drouillard <cdrllrd@google.com>