[debugger] Use new GetConcreteType calls.
Use the new GetConcreteTypeAs<> and the GetConcreteType variant that
takes a LazySymbol to clean up some call sites.

Remove the GetConcreteType "convenience wrapper" from ExprValue. It was
not actually more convenient, it just wrapped the call with the opposite
order of evaluation and didn't support the new variants.

Change a couple of symbol "AsFoo" calls to the new "As<Foo>" but only
where I was already changing code. The rest of these will be converted
later.

Change-Id: Ia992da559d2b087301b3eb6c82d506441d2adb1f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/543141
Reviewed-by: Dangyi Liu <dangyi@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>