[debug] Update to allow multiple instances
Originally, before launching debug agent, the script checked for any
running instance of debug agent on target. Due to this, two instances of
debug agent on different ports could not be started with the script. The
check is now changed to look for open ports on the device.

Also fixed passing extra flags to zxdb.

Fixed: 78791
Change-Id: I0604404b1a5cf98cc805cb3456753e008f674ea4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/543621
Reviewed-by: Brett Wilson <brettw@google.com>
Commit-Queue: Puneetha Ramachandra <puneetha@google.com>