[reland][time] Remove the zx_clock_get syscall.
We had to rollback the original due to a prebuilt that was
using an SDK well outside the 6 week compatibility window
and so still contained a zx_clock_get reference
(ref b/189107475).

That problem is being resolved with the help of the DX team
so we should be good to reland after the long weekend.


This is a reland of 22c0ce0edeb4a1c328aad341cfca6d42e9661643

Original change's description:
> [time] Remove the zx_clock_get syscall.
>
> Following RFC-0008 we removed all the users of the kernel UTC clock in
> F2, so it should now be possible to finally remove the zx_clock_adjust
> and zx_clock_get syscalls. zx_clock_adjust was covered in fxr/515400.
>
> Bug: 74471
> Change-Id: I02579c0f44d21865b3e5b82d4e6d99de7d6eda16
> Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/516981
> Commit-Queue: Jody Sankey <jsankey@google.com>
> Reviewed-by: Roland McGrath <mcgrathr@google.com>
> Reviewed-by: Adam Barth <abarth@google.com>
> API-Review: Roland McGrath <mcgrathr@google.com>

Bug: 74471
Change-Id: Ia87c547f63802b01ebcf5aae3ce5fc6f88febab0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/535269
Commit-Queue: Jody Sankey <jsankey@google.com>
Reviewed-by: James Robinson <jamesr@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Reviewed-by: John Grossman <johngro@google.com>
API-Review: Roland McGrath <mcgrathr@google.com>